requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.6
django-extensions==3.1.3
django-widget-tweaks==1.4.8
Pillow==8.3.1
pkg_resources==0.0.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
django-filter==2.4.0